+++ cv_pipeliner/__init__.py
@@ -1,10 +1,10 @@
-__version__ = "0.9.0"
+__version__ = "0.10.0"
 
 from cv_pipeliner.core.data import ImageData, BboxData
 from cv_pipeliner.batch_generators.image_data import BatchGeneratorImageData
 from cv_pipeliner.batch_generators.bbox_data import BatchGeneratorBboxData
 from cv_pipeliner.visualizers.core.image_data import visualize_image_data, visualize_images_data_side_by_side
-from cv_pipeliner.metrics.image_data_matching import ImageDataMatching
+from cv_pipeliner.metrics.image_data_matching import ImageDataMatching, BboxDataMatching
 from cv_pipeliner.visualizers.core.image_data_matching import visualize_image_data_matching_side_by_side
 from cv_pipeliner.metrics.detection import get_df_detection_metrics, get_df_detection_recall_per_class
 from cv_pipeliner.metrics.classification import get_df_classification_metrics
@@ -13,13 +13,14 @@ from cv_pipeliner.utils.images_datas import (
     rotate_image_data, crop_image_data, get_perspective_matrix_for_base_keypoints,
     apply_perspective_transform_to_image_data,
     thumbnail_image_data, non_max_suppression_image_data, uncrop_bboxes_data,
-    resize_image_data
+    resize_image_data, split_image_by_grid, flatten_additional_bboxes_data_in_image_data
 )
 from cv_pipeliner.inference_models.detection.object_detection_api import (
     ObjectDetectionAPI_KFServing, ObjectDetectionAPI_ModelSpec, ObjectDetectionAPI_TFLite_ModelSpec,
     ObjectDetectionAPI_pb_ModelSpec
 )
 from cv_pipeliner.inference_models.detection.pytorch import PytorchDetection_ModelSpec
+from cv_pipeliner.inference_models.detection.yolov5 import YOLOv5_ModelSpec
 from cv_pipeliner.inference_models.classification.tensorflow import (
     TensorFlow_ClassificationModelSpec, TensorFlow_ClassificationModelSpec_TFServing
 )
@@ -27,3 +28,8 @@ from cv_pipeliner.inference_models.pipeline import PipelineModelSpec, PipelineMo
 from cv_pipeliner.inferencers.detection import DetectionInferencer
 from cv_pipeliner.inferencers.classification import ClassificationInferencer
 from cv_pipeliner.inferencers.pipeline import PipelineInferencer
+from cv_pipeliner.data_converters.coco import COCODataConverter
+from cv_pipeliner.data_converters.brickit import BrickitDataConverter
+from cv_pipeliner.data_converters.json import JSONDataConverter
+from cv_pipeliner.data_converters.supervisely import SuperviselyDataConverter
+from cv_pipeliner.utils.imagesize import get_image_size